FS-120 Split Tests Module
It turns out that most of the tests in the `tests` module only depend
on freestyle, and a few test on monix. We move tests to `freestyle`,
and we add monix as a test-only dependency.
